requirements.txt
django==5.1.7
djangorestframework==3.15.2
channels==4.2.0
gunicorn==21.2.0
redis==5.2.1
psycopg2-binary